Use numeric type for tokenId as required by schema
- Change from bigint to numeric (exact type schema expects)
- All variable types now match GraphQL schema requirements

+++ lib/for-wenbo-main/queries/market.ts
@@ -70,7 +70,7 @@ export const GET_MARKET_LIQUIDITY = gql`
 export const GET_MARKET_TOKEN_SUPPLY = gql`
   query GetMarketTokenSupply(
     $marketAddress: String!
-    $tokenId: bigint!
+    $tokenId: numeric!
     $limit: Int!
     $offset: Int!
   ) {
